Obtain imagery tiles

The ceramic package provides tools to download and load imagery and raster tiles from online servers.

Details

Any process that can trigger downloads will first check the ceramic_cache() in case the tile already exists.

The main functions are for downloading tiles and loading them as raster objects, and each accepts a spatial object for the first argument, alternatively a raster extent, or location:

get_tiles

Download tiles for a given service for an extent and resolution

get_tiles_buffer

Download tiles based on location and buffer (width, height) in metres

get_tiles_dim

Download tiles based on extent and output dimension in pixels

get_tiles_zoom

Download tiles base on extent and zoom level

Two helper functions will trigger the download of tiles and also collate the result into a raster object:

cc_location

Download tiles and build a raster object of imagery

cc_elevation

Download tiles and build a raster object of elevation data

Administration functions for handling the file cache and required API key for on online service:

get_api_key

Return the stored key for online API, or NULL

ceramic_cache

Report the location of the tile cache

clear_ceramic_cache

Delete all files in the tile cache (use with caution!)

Other functions that are either rarely used or considered subject to change:

ceramic_tiles

Find particular tiles from the cache

mercator_tile_extent

Abstract raster-extent form of the spherical Mercator tile system, expressed in tile-index and zoom

plot_tiles

Plot the tiles from

ceramic_tiles

tiles_to_polygon

Convert

ceramic_tiles	to simple features format

cc_casey

Specific location hardcoded form of

cc_location

cc_davis

Specific location hardcoded form of

cc_location

cc_heard

Specific location hardcoded form of

cc_location

cc_kingston

Specific location hardcoded form of

cc_location

cc_macquarie

Specific location hardcoded form of

cc_location

cc_mawson

Specific location hardcoded form of

cc_location